Broke classes and traits out into separate files.
