+++ src/main/scala/dpla/ingestion3/harvesters/oai/refactor/OaiProtocol.scala
@@ -1,0 +1,12 @@
+package dpla.ingestion3.harvesters.oai.refactor
+
+class OaiProtocol extends OaiMethods {
+
+  override def listAllRecordPages = ???
+
+  override def listAllRecordPagesForSet(set: String) = ???
+
+  override def parsePageIntoRecords(page: String) = ???
+
+  override def listAllSets = ???
+}
